+++ packages/components/avatar/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @jala-banyu/avatar
 
+## 6.0.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.8
+
+## 6.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.7
+
+## 6.0.7
+
+### Patch Changes
+
+- fix img component avatar
+
 ## 6.0.6
 
 ### Patch Changes