+++ packages/components/card/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jala-banyu/card
 
+## 6.0.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.8
+  - @jala-banyu/ripple@6.0.8
+
+## 6.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.7
+  - @jala-banyu/ripple@6.0.7
+
 ## 6.0.7
 
 ### Patch Changes