+++ packages/components/checkbox/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @jala-banyu/checkbox
 
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.8
+
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.7
+
 ## 3.0.6
 
 ### Patch Changes