+++ packages/components/image/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @jala-banyu/image
 
+## 6.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.8
+
+## 6.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.7
+
 ## 6.0.6
 
 ### Patch Changes