+++ packages/components/select/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @jala-banyu/select
 
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.8
+  - @jala-banyu/listbox@3.0.8
+  - @jala-banyu/popover@5.0.8
+  - @jala-banyu/scroll-shadow@3.0.8
+  - @jala-banyu/spinner@6.0.8
+
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.6.7
+  - @jala-banyu/listbox@3.0.7
+  - @jala-banyu/popover@5.0.7
+  - @jala-banyu/scroll-shadow@3.0.7
+  - @jala-banyu/spinner@6.0.7
+
 ## 3.0.6
 
 ### Patch Changes